requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
tzdata==2024.2
sqlparse==0.5.2
drf-yasg==1.21.8
python-dotenv==1.0.1